contol panel
The control panel is connected. Now the path is drawn only when the button is pressed. It is important that the pause is designed for a long pause of the user, so it does not clear the array of rendering points. For now the rendering is constant, in the future I’m thinking of adding a timer for drawing at intervals